Hotfix expiration timestamp format
The one used previously was not recognized by Firefox nor Safari. This is a quick fix but I need to think about better way to represent the expiration. I think I will just store the day when the solution is valid and handle the timezones and expiration in the app itself.